move identical code to separate file
