+++ src/gather/multi_threaded/benchmark_agg_avx_32bit.cpp
@@ -51,47 +51,7 @@ void print_max_tput( std::ostream& log, multithreaded_measures& results ) {
     log << "Cores: " << max_cores << " TPut: " << curr_max << std::endl;
 }
 
-void log_multithreaded_results( std::ostream& logfile, const size_t stride_size, const std::string ident, multithreaded_measures& results ) {
-    for ( auto it = results.begin(); it != results.end(); ++it ) {
-        logfile << ident << " " << stride_size << " " << stride_size * 8 << " " << it->first << " " << it->second.throughput << std::endl;
-    }
-}
-
-/* We anticipate the following order: scalar, linear, gather, seti */
-void log_multithreaded_results_per_file( std::string basename, const size_t stride_size, std::vector< multithreaded_measures* > results, bool clean ) {
-    std::vector< uint64_t > keys;
-    /* Extract all keys */
-    for ( auto it = results[0]->begin(); it != results[0]->end(); ++it ) {
-        keys.push_back( it->first );
-    }
-
-    if ( clean ) {
-        for ( auto key : keys ) {
-            std::string del_file = basename + "_" + std::to_string( key ) + "_cores.dat";
-            if ( remove( del_file.c_str() ) == 0 ) {
-                std::cout << "Succesfully removed " << del_file << " before the benchmark." << std::endl;
-            } else {
-                std::cout << "ERROR removing " << del_file << " before the benchmark (maybe file was not present anyway). CHECK RESULTS" << std::endl;
-            }
-        }
-    }
-
-    for ( auto key : keys ) {
-        std::ofstream out( basename + "_" + std::to_string( key ) + "_cores.dat", std::ios_base::app );
-        out << stride_size << " " << stride_size * 8;
-        for ( auto r : results ) {
-            out << " " << (*r)[ key ].mis << " " << (*r)[ key ].throughput;
-        }
-        out << std::endl;
-        out.close();
-    }
-}
-
-void print_multithreaded_results( std::ostream& logfile, std::string ident, multithreaded_measures& results ) {
-    for ( auto it = results.begin(); it != results.end(); ++it ) {
-        logfile << "[" << ident << "] Core Count: " << it->first << " TPut: " << it->second.throughput << std::endl;
-    }
-}
+#include "log_multithreaded_results.cpp"
 
 template <typename T>
 void generate_random_values(T* array, uint64_t number) {
@@ -294,11 +254,6 @@ int main(int argc, char** argv) {
         else {
             cout <<"Set - Stride with Size "<<stride_size<<" failed"<<endl;
         }
-        // writing results to file
-        // log_multithreaded_results( result_file, stride_size, "scalar", scalar ); /* Enable onle for single-file-logging! */
-        // log_multithreaded_results( result_file, stride_size, "linear", linear ); /* Enable onle for single-file-logging! */
-        // log_multithreaded_results( result_file, stride_size, "gather", gather ); /* Enable onle for single-file-logging! */
-        // log_multithreaded_results( result_file, stride_size, "seti", seti ); /* Enable onle for single-file-logging! */
         /* Write all to one file */
         log_multithreaded_results_per_file( "./data/256_32bits/results_" + std::to_string(p), stride_pow, { &scalar, &linear, &gather, &seti }, first_run );
         if ( first_run ) {
@@ -310,7 +265,6 @@ int main(int argc, char** argv) {
     print_multithreaded_results( cout, "gather", gather );
     print_multithreaded_results( cout, "seti", seti );
 
-    // result_file.close(); /* Enable onle for single-file-logging! */
     free(array_32);
 
     return 0;
