+++ src/gather/multi_threaded/benchmark_agg_avx_64bit.cpp
@@ -39,47 +39,8 @@ std::thread* create_thread( const uint64_t tid, uint64_t* local_result, double* 
     return t;
 }
 
-void log_multithreaded_results( std::ostream& logfile, const size_t stride_size, const std::string ident, multithreaded_measures& results ) {
-    for ( auto it = results.begin(); it != results.end(); ++it ) {
-        logfile << ident << " " << stride_size << " " << stride_size * 8 << " " << it->first << " " << it->second.throughput << std::endl;
-    }
-}
+#include "log_multithreaded_results.cpp"
 
-/* We anticipate the following order: scalar, linear, gather, seti */
-void log_multithreaded_results_per_file( std::string basename, const size_t stride_size, std::vector< multithreaded_measures* > results, bool clean ) {
-    std::vector< uint64_t > keys;
-    /* Extract all keys */
-    for ( auto it = results[0]->begin(); it != results[0]->end(); ++it ) {
-        keys.push_back( it->first );
-    }
-
-    if ( clean ) {
-        for ( auto key : keys ) {
-            std::string del_file = basename + "_" + std::to_string( key ) + "_cores.dat";
-            if ( remove( del_file.c_str() ) == 0 ) {
-                std::cout << "Succesfully removed " << del_file << " before the benchmark." << std::endl;
-            } else {
-                std::cout << "ERROR removing " << del_file << " before the benchmark (maybe file was not present anyway). CHECK RESULTS" << std::endl;
-            }
-        }
-    }
-
-    for ( auto key : keys ) {
-        std::ofstream out( basename + "_" + std::to_string( key ) + "_cores.dat", std::ios_base::app );
-        out << stride_size << " " << stride_size * 8;
-        for ( auto r : results ) {
-            out << " " << (*r)[ key ].mis << " " << (*r)[ key ].throughput;
-        }
-        out << std::endl;
-        out.close();
-    }
-}
-
-void print_multithreaded_results( std::ostream& logfile, std::string ident, multithreaded_measures& results ) {
-    for ( auto it = results.begin(); it != results.end(); ++it ) {
-        logfile << "[" << ident << "] Core Count: " << it->first << " TPut: " << it->second.throughput << std::endl;
-    }
-}
 template <typename T>
 void generate_random_values(T* array, uint64_t number) {
   static_assert(is_integral<T>::value, "Data type is not integral.");
@@ -274,11 +235,6 @@ int main(int argc, char** argv) {
             cout <<"Set - Stride with Size "<<stride_size<<" failed"<<endl;
         }
 
-        // writing results to file
-        // log_multithreaded_results( result_file, stride_pow, "scalar", scalar );  /* Enable onle for single-file-logging! */
-        // log_multithreaded_results( result_file, stride_pow, "linear", linear );  /* Enable onle for single-file-logging! */
-        // log_multithreaded_results( result_file, stride_pow, "gather", gather );  /* Enable onle for single-file-logging! */
-        // log_multithreaded_results( result_file, stride_pow, "seti", seti );  /* Enable onle for single-file-logging! */
         /* Write all to one file */
         log_multithreaded_results_per_file( "./data/256_64bits/results_" + std::to_string(p), stride_pow, { &scalar, &linear, &gather, &seti }, first_run );
         if ( first_run ) {
@@ -289,8 +245,6 @@ int main(int argc, char** argv) {
     print_multithreaded_results( cout, "linear", linear );
     print_multithreaded_results( cout, "gather", gather );
     print_multithreaded_results( cout, "seti", seti );
-
-    // result_file.close(); /* Enable onle for single-file-logging! */
 
     free(array_64);
     return 0;
